Android: APK imzolash, token yangilash, click-to-call va telefonga chiqish yo'li
Imzolash keystore.properties dan yoki muhitdan o'qiladi, v1+v2+v3 bilan.
Keystore bo'lmasa ham release quriladi (imzosiz) — CI va kaliti yo'q odam
ham release variantining kompilyatsiyasini tekshira oladi.

Dev host build vaqtida almashtiriladi: sinovchi local.properties ga o'z
IP sini yozadi, ya'ni hech kimning uy manzili repoga tushmaydi va hech kim
kuzatiladigan faylni tahrirlamaydi. Host ko'rsatilmasa, qator umuman
chiqarilmaydi — o'rniga qolgan placeholder hech kim o'qimagan konfiguratsiya.

Token: bironta holat navbatni tashlamaydi. Testda enum bo'ylab yuriladi,
ya'ni keyin qo'shilgan beshinchi holat ham buni jimgina buza olmaydi.
Va faqat REFRESH rad etilgani auth_expired beradi — bitta 401 shunchaki
eskirgan token, refresh paytidagi tarmoq uzilishi esa umuman muddat emas:
sessiya tirik bo'lishi mumkin, telefon shunchaki liftda.

UPDATE_REQUIRED holatida ilova YUBORISHDA DAVOM ETADI. Server eski
klientning navbatini avval qabul qiladi va faqat bo'shagach rad etadi —
to'xtagan klient aynan darvoza himoya qilayotgan ma'lumotni qamab qo'yardi.

Click-to-call: eskirgan buyruq jimgina tashlanmaydi, "eskirgan" deb
tasdiqlanadi — panel "kech keldi" bilan "umuman kelmadi" ni ajrata olishi
kerak, bular turli nosozlik va turli yechim.

int64 vaqtinchalik skripti o'zini o'chirdi: backend maydonlarga format
qo'ygach test qizardi, bu esa "endi kerak emas" signali edi.

docs/ON-DEVICE-TESTING.md — noutbukdan ushlangan qo'ng'iroqqacha sakkiz qadam.
docs/APK-SIGNING.md — kalit zaxirasi va uni yo'qotish nimaga turishi.

+++ server/src/modules/audio/reports.py
@@ -7,24 +7,26 @@ from datetime import date
 
 from pydantic import BaseModel, Field
 
+from src.core.wire import Int64
+
 
 class StoragePointOut(BaseModel):
     """One day of the growth curve."""
 
     period_date: date
-    audio_bytes_total: int
+    audio_bytes_total: Int64
     audio_files: int
-    bytes_added: int
-    bytes_deleted: int
+    bytes_added: Int64
+    bytes_deleted: Int64
 
 
 class StorageReportResponse(BaseModel):
     """N18: current usage, 30-day growth, and the projection it implies."""
 
-    audio_bytes_total: int
+    audio_bytes_total: Int64
     audio_files: int
-    bytes_added_30d: int
-    projected_bytes_12m: int = Field(
+    bytes_added_30d: Int64
+    projected_bytes_12m: Int64 = Field(
         description=(
             "Today's total plus a year at the last 30 days' rate. The provision "
             "is 250 GB, and this is what says whether that is enough."
@@ -38,10 +40,10 @@ class DataUsageRowOut(BaseModel):
 
     installation_id: uuid.UUID
     agent_name: str
-    cellular_bytes_month: int
-    wifi_bytes_month: int
+    cellular_bytes_month: Int64
+    wifi_bytes_month: Int64
     requests_month: int
-    cap_bytes_month: int
+    cap_bytes_month: Int64
     over_cap: bool = Field(
         description="Past N14's cap: the app stops uploading audio over cellular."
     )
@@ -50,4 +52,4 @@ class DataUsageRowOut(BaseModel):
 class DataUsageResponse(BaseModel):
     items: list[DataUsageRowOut]
     total: int
-    cap_bytes_month: int
+    cap_bytes_month: Int64
